Chatbots/requirements.txt
langchain
streamlit
openai
python-dotenv
langchain-core
langchain-google-genai
langchain_community
flask
